Payment fixes

+++ mvp/views.py
@@ -328,34 +328,59 @@ def chat(request, user_id=None):
         return render(request, './chat.html', {
             'messages': messages,
             'contact': contact
-    })
+        })
 
 
 @login_required(login_url='/login')
 def paypal(request, contact_id):
-    offer_list = None
+    offer_list = []
     try:
         contact = get_object_or_404(Venue, pk=contact_id)
-        offer_list = Offer.objects.filter(
-            venue_id=contact.id).order_by('-date')
     except:
         contact = get_object_or_404(Artist, pk=contact_id)
+        offer_list = Offer.objects.filter(
+            venue_id=request.user.id).order_by('-date')
 
     principal = request.user
 
-    if (offer_list is None):
-        offer_list = Offer.objects.filter(
-            venue_id=principal.id).order_by('-date')
+    paymentErrors = []
 
-    context = {'contact': contact, 'user': principal, 'offer_list': offer_list}
+    if ('paymentErrors' in request.session):
+        paymentErrors = request.session['paymentErrors']
+        request.session['paymentErrors'] = []
+
+    context = {'contact': contact, 'user': principal,
+               'offer_list': offer_list, 'paymentErrors': paymentErrors}
     return render(request, './paypal.html', context)
 
 
 @login_required(login_url='/login')
 def payment(request):
 
+    errors = []
+
     form = request.POST
     payee = get_object_or_404(User, id=form['payee'])
+
+    # Aquí se puede añadir validación adicional sobre el destinatario
+    if (payee is None or payee.id == request.user.id):
+        errors.append("Destinatario del pago no válido.")
+
+    if (form['amount'] is None or form['amount'] == '' or float(form['amount']) <= 0.0):
+        errors.append(
+            "Por favor, introduce una cantidad entre 1 y 9999 euros.")
+
+    if (form['performanceDate'] is None or form['performanceDate'] == ""):
+        errors.append("Por favor, introduzca una fecha.")
+
+    dateFormat = "%Y-%m-%d"
+
+    try:
+        formDate = datetime.strptime(form['performanceDate'], dateFormat)
+        if (formDate.date() < datetime.today().date()):
+            errors.append("Por favor, introduzca una fecha en el futuro.")
+    except:
+        errors.append("Ha ocurrido un error procesando la fecha.")
 
     try:
         venue = Venue.objects.get(id=request.user.id).id
@@ -372,6 +397,10 @@ def payment(request):
         'artist': artist,
         'venue': venue
     }
+
+    if errors:
+        request.session['paymentErrors'] = errors
+        return JsonResponse({'errors': errors, 'status': 'formError'})
 
     print('============ Performance info: ============')
 
@@ -576,14 +605,14 @@ def payout(request):
     paymentObject.amount = payment['transactions'][0]['amount']['total']
     paymentObject.user = request.user
     paymentObject.performance = performance
-    paymentObject.date = datetime.datetime.now()
+    paymentObject.date = datetime.now()
     paymentObject.paypalId = payment['id']
 
     paymentObject.save()
     print('============ Payment saved ============')
 
     offerId = request.session['relaterOffer']
-    if (offerId != 0):
+    if (offerId is not None and offerId != 'undefined' and offerId != 0):
         offer = Offer.objects.get(id=offerId)
         offer.delete()
 
@@ -594,8 +623,6 @@ def payout(request):
 def paymentConfirmation(request):
     payment = request.session['payment']
     return render(request, './paypalConfirm.html', {'payment': payment})
-
-
 
 
 class formulario_feedback(View):
